move backed up files
