+++ bak/tests-bak/data_files/camelcase.stream
