+++ bak/tests-bak/data_files/camelcase_complex_schema.stream
