+++ bak/tests-bak/data_files/encoded_strings.stream
