+++ bak/tests-bak/data_files/record_before_schema.stream
