+++ bak/tests-bak/data_files/schema_no_properties.stream
