+++ bak/tests-bak/data_files/special_chars_in_attributes.stream
